chore: rename store files and remove unused test infrastructure
Made-with: Cursor
