+++ components/icon-page/panels/workspace/hooks/use-workspace-selection.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { useWorkspaceSelectionStore } from "@/stores/use-workspace-selection-store";
+import { useWorkspaceSelectionStore } from "@/stores/workspace-selection";
 
 export function useWorkspaceSelection() {
   return useWorkspaceSelectionStore();
