+++ package.json
@@ -45,25 +45,19 @@
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.2.2",
-    "@testing-library/dom": "^10.4.1",
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.2",
     "@trivago/prettier-plugin-sort-imports": "^6.0.2",
     "@types/dompurify": "^3.2.0",
     "@types/node": "^25",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "@vitejs/plugin-react": "^6.0.1",
     "babel-plugin-react-compiler": "1.0.0",
     "eslint": "^10.2.0",
     "eslint-config-next": "16.2.2",
-    "jsdom": "^29.0.1",
     "postcss": "^8.5",
     "prettier": "^3.8.1",
     "prettier-plugin-tailwindcss": "^0.7.2",
     "tailwindcss": "^4.2.2",
     "tw-animate-css": "1.4.0",
-    "typescript": "^6",
-    "vitest": "^4.1.2"
+    "typescript": "^6"
   }
 }
