+++ stores/editor-saved-assets.ts
