+++ stores/editor-selection.ts
